Add CreateShardingTableRuleStatementUpdater
