+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.RDLUpdater
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-
+org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateShardingTableRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingTableRuleStatementUpdater
